Refactor: Extract JSON utils to dedicated module
